Modules

Grids

Chmy.Grids.StructuredGrid — Type

StructuredGrid

Represents a structured grid with orthogonal axes.

Chmy.Grids.UniformGrid — Method

UniformGrid(arch; origin, extent, dims, topology=nothing)

Constructs a uniform grid with specified origin, extent, dimensions, and topology.

Arguments

arch::Architecture: The associated architecture.

origin::NTuple{N,Number}: The origin of the grid.

extent::NTuple{N,Number}: The extent of the grid.

dims::NTuple{N,Integer}: The dimensions of the grid.

topology=nothing: The topology of the grid. If not provided, a default Bounded topology is used.

Chmy.Grids.axis — Method

axis(grid, dim::Dim)

Return the axis corresponding to the spatial dimension dim.

Chmy.Grids.bounds — Method

bounds(grid, loc, [dim::Dim])

Return the bounds of a structured grid at the specified location(s).

Chmy.Grids.connectivity — Method

connectivity(grid, dim::Dim, side::Side)

Return the connectivity of the structured grid grid for the given dimension dim and side side.

Chmy.Grids.coord — Method

coord(grid, loc, I...)

Return a tuple of spatial coordinates of a grid point at location loc and indices I.

For vertex locations, first grid point is at the origin. For center locations, first grid point at half-spacing distance from the origin.

Chmy.Grids.extent — Method

extent(grid, loc, [dim::Dim])

Return the extent of a structured grid at the specified location(s).

Chmy.Grids.inv_spacing — Method

inv_spacing(grid, loc, I...)

Return a tuple of inverse grid spacings at location loc and indices I.

Chmy.Grids.origin — Method

origin(grid, loc, [dim::Dim])

Return the origin of a structured grid at the specified location(s).

Chmy.Grids.spacing — Method

spacing(grid, loc, I...)

Return a tuple of grid spacings at location loc and indices I.

Architectures

Chmy.Architectures.Architecture — Type

Architecture

Abstract type representing an architecture.

Chmy.Architectures.SingleDeviceArchitecture — Type

SingleDeviceArchitecture <: Architecture

A struct representing an architecture that operates on a single CPU or GPU device.

Chmy.Architectures.Arch — Method

Arch(backend::Backend; device_id::Integer=1)

Create an architecture object for the specified backend and device.

Arguments

backend: The backend to use for computation.

device_id=1: The ID of the device to use.

Chmy.Architectures.activate! — Method

activate!(arch::SingleDeviceArchitecture; priority=:normal)

Activate the given architecture on the specified device and set the priority of the backend.

Chmy.Architectures.get_backend — Method

get_backend(arch::SingleDeviceArchitecture)

Get the backend associated with a SingleDeviceArchitecture.

Chmy.Architectures.get_device — Method

get_device(arch::SingleDeviceArchitecture)

Get the device associated with a SingleDeviceArchitecture.

Fields

Chmy.Fields.ConstantField — Type

Scalar field with a constant value

Chmy.Fields.Field — Type

struct Field{T,N,L,H,A} <: AbstractField{T,N,L}

Field represents a discrete scalar field with specified type, number of dimensions, location, and halo size.

Chmy.Fields.Field — Method

Field(backend, grid, loc, type=eltype(grid); halo=1)

Constructs a field on a structured grid at the specified location.

Arguments:

backend: The backend to use for memory allocation.

grid: The structured grid on which the field is constructed.

loc: The location or locations on the grid where the field is constructed.

type: The element type of the field. Defaults to the element type of the grid.

halo: The halo size for the field. Defaults to 1.

Chmy.Fields.FunctionField — Type

FunctionField <: AbstractField

Continuous or discrete field with values computed at runtime.

Constructors

FunctionField(func, grid, loc; [discrete], [parameters]): Create a new FunctionField object.

Chmy.Fields.OneField — Type

Constant field with values equal to one(T)

Chmy.Fields.ValueField — Type

Field with a constant value

Chmy.Fields.ZeroField — Type

Constant field with values equal to zero(T)

Grid Operators

Chmy.GridOperators.leftx — Method

leftx(f, loc, I)

"left side" of a field ([1:end-1]) in x direction.

Chmy.GridOperators.lefty — Method

lefty(f, loc, I)

"left side" of a field ([1:end-1]) in y direction.

Chmy.GridOperators.leftz — Method

leftz(f, loc, I)

"left side" of a field ([1:end-1]) in z direction.

Chmy.GridOperators.rightx — Method

rightx(f, loc, I)

"right side" of a field ([2:end]) in x direction.

Chmy.GridOperators.righty — Method

righty(f, loc, I)

"right side" of a field ([2:end]) in y direction.

Chmy.GridOperators.rightz — Method

rightz(f, loc, I)

"right side" of a field ([2:end]) in z direction.

Chmy.GridOperators.δx — Method

δx(f, loc, I)

Finite difference in x direction.

Chmy.GridOperators.δy — Method

δy(f, loc, I)

Finite difference in y direction.

Chmy.GridOperators.δz — Method

δz(f, loc, I)

Finite difference in z direction.

Chmy.GridOperators.∂x — Method

∂x(f, loc, grid, I)

Directional partial derivative in x direction.

Chmy.GridOperators.∂y — Method

∂y(f, loc, grid, I)

Directional partial derivative in y direction.

Chmy.GridOperators.∂z — Method

∂z(f, loc, grid, I)

Directional partial derivative in z direction.

Boundary Conditions

Chmy.BoundaryConditions.AbstractBatch — Type

AbstractBatch

Abstract type representing a batch of boundary conditions.

Chmy.BoundaryConditions.EmptyBatch — Type

EmptyBatch <: AbstractBatch

EmptyBatch represents no boundary conditions.

Chmy.BoundaryConditions.ExchangeBatch — Type

ExchangeBatch <: AbstractBatch

ExchangeBatch represents a batch used for MPI communication.

Chmy.BoundaryConditions.FieldBatch — Type

FieldBatch <: AbstractBatch

FieldBatch is a batch of boundary conditions, where each field has one boundary condition.

Chmy.BoundaryConditions.FieldBoundaryCondition — Type

FieldBoundaryCondition

Abstract supertype for all boundary conditions that are specified per-field.

Chmy.BoundaryConditions.bc! — Method

bc!(arch::Architecture, grid::StructuredGrid, batch::BatchSet)

Apply boundary conditions using a batch set batch containing an AbstractBatch per dimension per side of grid.

Arguments

arch: The architecture.

grid: The grid.

batch:: The batch set to apply boundary conditions to.

Kernel launcher

Distributed

Chmy.Distributed.CartesianTopology — Type

CartesianTopology

Represents N-dimensional Cartesian topology of distributed processes.

Chmy.Distributed.CartesianTopology — Method

CartesianTopology(comm, dims)

Create an N-dimensional Cartesian topology using base MPI communicator comm with dimensions dims. If all entries in dims are not equal to 0, the product of dims should be equal to the total number of MPI processes MPI.Comm_size(comm). If any (or all) entries of dims are 0, the dimensions in the corresponding spatial directions will be picked automatically.

Chmy.Distributed.DistributedArchitecture — Type

DistributedArchitecture <: Architecture

A struct representing a distributed architecture.

Chmy.Distributed.StackAllocator — Type

mutable struct StackAllocator

Simple stack (a.k.a. bump/arena) allocator. Maintains an internal buffer that grows dynamically if the requested allocation exceeds current buffer size.

Chmy.Distributed.StackAllocator — Method

StackAllocator(backend::Backend)

Create a stack allocator using the specified backend to store allocations.

Base.resize! — Method

resize!(sa::StackAllocator, sz::Integer)

Resize the StackAllocator's buffer to capacity of sz bytes. This method will throw an error if any arrays were already allocated using this allocator.

Chmy.Architectures.Arch — Method

Arch(backend, comm, dims; kwargs...)

Create a distributed Architecture using backend backend and comm. For GPU backends, device will be selected automatically based on a process id within a node.

Chmy.Distributed.allocate — Function

allocate(sa::StackAllocator, T::DataType, dims, [align=sizeof(T)])

Allocate a buffer of type T with dimensions dims using a stack allocator. The align parameter specifies the alignment of the buffer elements.

Arguments

sa::StackAllocator: The stack allocator object.

T::DataType: The data type of the requested allocation.

dims: The dimensions of the requested allocation.

align::Integer: The alignment of the allocated buffer in bytes.

Warning

Arrays allocated with StackAllocator are not managed by Julia runtime. User is responsible for ensuring correct lifetimes, i.e., that the reference to allocator outlives all arrays allocated using this allocator.

Chmy.Distributed.cart_comm — Method

cart_comm(topo)

MPI Cartesian communicator for the topology.

Chmy.Distributed.coords — Method

coords(topo)

Coordinates of a current process within a Cartesian topology.

Chmy.Distributed.dims — Method

dims(topo)

Dimensions of the topology as NTuple.

Chmy.Distributed.exchange_halo! — Method

exchange_halo!(side::Side, dim::Dim, arch, grid, fields...; async=false)

Perform halo exchange communication between neighboring processes in a distributed architecture.

Arguments

side: The side of the grid where the halo exchange is performed.

dim: The dimension along which the halo exchange is performed.

arch: The distributed architecture used for communication.

grid: The structured grid on which the halo exchange is performed.

fields...: The fields to be exchanged.

Optional Arguments

async=false: Whether to perform the halo exchange asynchronously.

Chmy.Distributed.exchange_halo! — Method

exchange_halo!(arch, grid, fields...)

Perform halo exchange for the given architecture, grid, and fields.

Arguments

arch: The distributed architecture to perform halo exchange on.

grid: The structured grid on which halo exchange is performed.

fields: The fields on which halo exchange is performed.

Chmy.Distributed.gather! — Method

gather!(arch, dst, src::Field; kwargs...)

Gather the interior of a field src into a global array dst.

Chmy.Distributed.gather! — Method

gather!(dst, src, comm::MPI.Comm; root=0)

Gather local array src into a global array dst. Size of the global array size(dst) should be equal to the product of the size of a local array size(src) and the dimensions of a Cartesian communicator comm. The array will be gathered on the process with id root (root=0 by default). Note that the memory for a global array should be allocated only on the process with id root, on other processes dst can be set to nothing.

Chmy.Distributed.global_rank — Method

global_rank(topo)

Global id of a process in a Cartesian topology.

Chmy.Distributed.global_size — Method

global_size(topo)

Total number of processes withing the topology.

Chmy.Distributed.has_neighbor — Method

has_neighbor(topo, dim, side)

Returns true if there a neighbor process in spatial direction dim on the side side, or false otherwise.

Chmy.Distributed.nallocs — Method

nallocs(sa::StackAllocator)

Get the number of allocations made by the given StackAllocator.

Chmy.Distributed.neighbor — Method

neighbor(topo, dim, side)

Returns id of a neighbor process in spatial direction dim on the side side, if this neighbor exists, or MPI.PROC_NULL otherwise.

Chmy.Distributed.neighbors — Method

neighbors(topo)

Neighbors of a current process.

Returns tuple of ranks of the two immediate neighbors in each spatial direction, or MPI.PROC_NULL if there is no neighbor on a corresponding side.

Chmy.Distributed.node_name — Method

node_name(topo)

Name of a node according to MPI.Get_processor_name().

Chmy.Distributed.node_size — Method

node_size(topo)

Number of processes sharing the same node.

Chmy.Distributed.reset! — Method

reset!(sa::StackAllocator)

Reset the stack allocator by resetting the pointer. Doesn't free the internal memory buffer.

Chmy.Distributed.shared_comm — Method

shared_comm(topo)

MPI communicator for the processes sharing the same node.

Chmy.Distributed.shared_rank — Method

shared_rank(topo)

Local id of a process within a single node. Can be used to set the GPU device.

Workers

Chmy.Workers.Worker — Type

Worker

A worker that performs tasks asynchronously.

Constructor

Worker{T}(; [setup], [teardown]) where {T}

Constructs a new Worker object.

Arguments

setup: A function to be executed before the worker starts processing tasks. (optional)

teardown: A function to be executed after the worker finishes processing tasks. (optional)